Rename address-lookup folder
To keep with convention we name the folder after the service we are sending requests to, rather than what it does.
